+++ projects/ui-components/internal-shared/validation-utils.spec.ts
@@ -17,4 +17,24 @@ describe('ValidationUtils', () => {
     expect(badFC.hasError('twoDigit')).toBeTruthy();
   });
 
+  it('should create requiredValidator that tests if the input is empty spaces', () => {
+
+    const validator = ValidationUtils.requiredValidator();
+    let formControl: FormControl = new FormControl('', [validator]);
+    expect(formControl.hasError('required')).toBeTrue();
+
+    formControl = new FormControl('  ', [validator]);
+    expect(formControl.hasError('required')).toBeTrue();
+
+    formControl = new FormControl(null, [validator]);
+    expect(formControl.hasError('required')).toBeTrue();
+
+    formControl = new FormControl(undefined, [validator]);
+    expect(formControl.hasError('required')).toBeTrue();
+
+    formControl = new FormControl('123', [validator]);
+    expect(formControl.hasError('required')).toBeFalse();
+
+  });
+
 });