+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 14.1.5
+
+### Features
+
+### Fixes
+  - Fixed bug where Required input field was not recognizing the character "s" as a valid input because of Regex issues. 
+    Changed to use LfMetadataValidationUtils.isNullOrEmpty to in the requiredValidator function in [validation-utils.ts](projects\ui-components\internal-shared\validation-utils.ts) instead of Regex.
+### Chore & Maintenance
+  - Updated `@laserfiche/lf-js-utils` to version `4.0.8` to use newly added LfMetadataValidationUtils.isNullOrEmpty for field validation.
+
 ## 14.1.4
 
 ### Features